requirements.txt
PyMuPDF
sentence-transformers
faiss-cpu
nltk
numpy
requests
torch
tqdm